Move Reference to vim25/mo
